refactor(fs): split passthrough filesystem
